# perf(browser): use `while true` loop