before: 

```
  meros                     x 2,865 ops/sec ±2.12% (61 runs sampled)
  fetch-multipart-graphql   x 2,656 ops/sec ±1.13% (65 runs sampled)
```

after:

```
  meros                     x 3,266 ops/sec ±0.64% (56 runs sampled)
  fetch-multipart-graphql   x 2,566 ops/sec ±1.53% (66 runs sampled)
```