@lukeed 2020-11-10
`TextDecoder` doesn't exist in Node 10.x